requirements.txt
langchain
langchain-groq
faiss-cpu
tiktoken
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-dotenv
pymupdf
unstructured
pdfminer.six
python-docx
docx2txt
pandas
openpyxl
jq
matplotlib
langchain-huggingface
langchain-experimental
rank_bm25
wikipedia
langgraph